Bibliotheque/src/test/javascript/spec/app/entities/jeux-educatif/jeux-educatif.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { JeuxEducatifService } from 'app/entities/jeux-educatif/jeux-educatif.service';
import { IJeuxEducatif, JeuxEducatif } from 'app/shared/model/jeux-educatif.model';

describe('Service Tests', () => {
  describe('JeuxEducatif Service', () => {
    let injector: TestBed;
    let service: JeuxEducatifService;
    let httpMock: HttpTestingController;
    let elemDefault: IJeuxEducatif;
    let expectedResult: IJeuxEducatif | IJeuxEducatif[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(JeuxEducatifService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new JeuxEducatif(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a JeuxEducatif', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new JeuxEducatif()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a JeuxEducatif', () => {
        const returnedFromService = Object.assign(
          {
            nom: 'BBBBBB',
            prix: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of JeuxEducatif', () => {
        const returnedFromService = Object.assign(
          {
            nom: 'BBBBBB',
            prix: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a JeuxEducatif', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
